Refactor API package structure and update import paths
