+++ pkg/platforms/windows/platform.go
@@ -2,7 +2,7 @@ package windows
 
 import (
 	"errors"
-	"github.com/wizzomafizzo/tapto/pkg/service/api/notifications"
+	"github.com/wizzomafizzo/tapto/pkg/api/notifications"
 	"os"
 	"os/exec"
 	"path/filepath"
